+++ bm-runner/bm_process.py
@@ -6,7 +6,7 @@ import resource
 import sys
 import subprocess
 from bm_executer import Executer
-from bm_executer import ExecutionUnit
+from bm_exec_unit import ExecutionUnit
 from bm_utils import stop_process
 from bm_config import Application
 from config.benchmark import ExecutionType
